PTI-536: Code review updates - Move business logic to business-logic-manager.js - Update instances of 'nris' to 'nris-epd', where appropriate - Update swagger enums
