+++ api/src/swagger/swagger.yaml
@@ -52,13 +52,26 @@ definitions:
     example: 'epic'
     enum: &dataSourceType
       - epic
+      - nris-epd
   recordType:
     type: string
     description: 'Supported record types'
-    example: 'order'
+    example: 'Order'
     enum: &recordType
-      - Order
+      - AdministrativePenalty
+      - AdministrativeSanction
+      - Agreement
+      - Certificate
+      - ConstructionPlan
+      - CourtConviction
       - Inspection
+      - ManagementPlan
+      - Order
+      - Permit
+      - RestorativeJustice
+      - SelfReport
+      - Ticket
+      - Warning
 
   ### Document Type Definitions
   DocumentRequestObject:
@@ -82,8 +95,20 @@ definitions:
     description: 'Optional datasets to search on in the DB'
     example: ['Order']
     enum: &searchDataSetParams
-      - Order
+      - AdministrativePenalty
+      - AdministrativeSanction
+      - Agreement
+      - Certificate
+      - ConstructionPlan
+      - CourtConviction
       - Inspection
+      - ManagementPlan
+      - Order
+      - Permit
+      - RestorativeJustice
+      - SelfReport
+      - Ticket
+      - Warning
   SearchSubsetParams:
     type: string
     description: 'Optional datasets to search on in the DB'
